+++ eslint.config.mjs
@@ -19,8 +19,7 @@ const eslintConfig = defineConfig([
       "@typescript-eslint/no-unused-vars": "off",
       "@typescript-eslint/no-unused-expressions": "off",
       "react-hooks/purity": "off",
-      "react-hooks/set-state-in-effect": "off",
-      "@next/next/no-img-element": "off"
+      "react-hooks/set-state-in-effect": "off"
     }
   }
 ]);